Update rpm spec file
This commit updates the rpm spec file to change
the ruby dependencies to version 1.8.5. It also
removes any conditionals for ruby abi, because
we can now assume ruby abi present. The
dependency on ruby is bumped to version 1.8.5,
as we can also assume this ruby version, given
that el4 is EOL. The facter version is bumped
to 1.6.8 to reflect the latest released version.
Finally, this commit removes the manual noarch
build setting, as facter is now arch-specific
because of its dependency on dmidecode.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>