(#7484) Domain fact should handle only TLD
Prior to this commit, the domain fact executed "hostname" on all platforms,
however the default on Linux is that "hostname" does not return FQDN.
This causes problems when the machine only has a TLD, rather than being
in a second level domain. For example, ns01.tld vs ns01.puppetlabs.com.

This commit fixes that issue, and handles the case where hostname does
not accept the -f option on Solaris.