(#6955) Remove relative dirs from fact search path
Prior to this commit, relative directories were permitted the search
path for facts, which meant that both the user and the path in which
facter was run could influence which facts were located.  This situation
could have led in unintended code being executed.

This commit forces all paths (from $LOAD_PATH, ENV["FACTORLIB"] or
Facter.search_path) to explicitly be absolute paths.