Partial revert of CFPropertyList migration.
While the new CFPropertyList implementation is great, the older plist
implementation is used from a bunch of places in Puppet - including some
package providers and other unexpected sources.

Since the two projects share this code so intimately, we risk breaking more
than we care about if we just eliminate this code.

This restores the previous version of those files, allowing dependent code to
gradually be migrated away to the newer implementation over time.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>