(maint) Fix broken Windows tests
One of the test was not using the correct path seperator on Windows,
so the : was changed to File::PATH_SEPARATOR, so the tests actually
test the right behavior on each platform.

Remove the test to see if " " is an absolute directory. It is
relative on Unix but seems to be absolute on Windows (both in
1.8.7 and 1.9.3). Considering this platform discrepency and the
fact that there is not valid use case for the " " directory, the
test is not nessecary.