(#14467) Warn when removing relative paths
Due to changes made to Facter in commit
634f2f6a57b461926afcb2e07dcfe4ed659f5b0c, Facter no longer looks
for facts in relative directories for security reasons. This
change could result in users not getting all of their facts. This
commit alerts users if a relative directory has been excluded from
the search path.

Due to feedback on this pull request on IRC, change message to be
debug instead of warn.