Use git describe in Rakefile to determine pkg ver
This commit borrows from the puppet-dashboard tarball
generation rake task, using git describe to generate
the version string instead of using the hardcoded
value in lib/facter.rb. This allows the tarball task
to generate correctly named RC tarballs. Currently,
packaging RCs requires the manual renaming and tarring
of the generated directory. Because the built in rake/
PackageTask does not play well with long git-describe
strings, it is replaced with a different implementation
that can handle the longer git-describe version strings.
The gem package task still uses the rake built-in, so
the version string is santized for it.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>