(#14764) Stub architecture fact when Windows facts run on Linux
Previously the Windows specific tests were failing when run on Linux
(but not Mac or Windows), because the processor fact always runs Linux,
Aix, and Solaris specific commands, causing the architecture fact to be
loaded, which loads the hardwaremodel, which was recently changed to use
WMI to retrieve the processor architecture instead of using RbConfig.

This commit just stubs the architecture fact until the processor fact
can be fixed to conditionally execute those commands based on the
current kernel fact.