(#14689) Fix SSH fact for HP-UX
Prior to this commit, mmoll had added the location `/etc/opt/ssh` to the
search path for the SSH related facts. This commit adds tests for the
additional directory in the search path.

Also prior to this commit, the SSH fact frighteningly re-defined the
fact resolutions for each directory within the search path. This commit
refactors the SSH facts to embed the search path checking inside the
individual facts rather than a re-definition of the fact for each
directory.