(#14827) Fix broken swap units on Mac
Prior to this commit, the swap values for Mac OS X were based solely on
the text returned from 'sysctl swapusage', which was something like
'6144.00M'. This resulting string was passed directly on to Facter
without scaling or transformation.  This commit makes the swap facts on
OS X consistent with the swap facts on every other OS in that the facts
will be dynamically scaled and in the same format: '6.14 GB'.