Bump Facter epoch to 1
This commit bumps the facter epoch to 1. This
is to address the errant release of a facter 2.0rc
to the Puppet Labs yum production repository, which
may have been then installed unintentionally by its
users. The epoch bump is in tandem with equivalent
changes to Facter 1.6.9 and 2.0.0rc4. This will
ensure users who intentionally installed 1.6.10 rc/
2.0 rc keep it, while users of 1.6.9 who unintentionally
installed 2.0rc are taken back to 1.6.9.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>