(#15001) Change Linux confined ipaddress fact to use 'ip addr' This commit changes the version of the ipaddress fact that is confined with Linux to use the 'ip' command.Recent versions of net-tool have stated that ifconfig is obsolete. It also states to rely on 'ip addr' as a replacement. The output was also changed to be more like that of the 'ip addr' output.
This commit also changes the fact to use
'Facter::Util::Resolution.exec'. The 'ip' command is left unqualified since the recent
changes to 'Facter::Util::Resolution.exec' include functionality that mimics the 'which'
command but does not need to fork. The switch to 'Facter::Util::Resolution.exec'
also allows for unit tests, which are part of this commit.