(#2066) Fix facts for Linux/kFreeBSD and add tests
Make swap and memory facts in MB for Linux/kFreeBSD, as was previously
done for other platforms. Also make scaled facts depend on MB versions.
This should fix the problems encountered in the Jenkins test of the
previously-submitted code. Add tests for memory and swap facts for Linux
and kFreeBSD.