(#15049) Return only one selinuxfs path as string from mounts
The block that parses /proc/self/mountinfo to find a selinuxfs filesystem would
return results as an array.  On Ruby 1.8, interpolating this into a string for
File.exists? when one result was returned worked, while on Ruby 1.9 it
interpolated as ["/sys/fs/selinux"]/enforce so later failed.

This changes the block to return the single result string rather than an array.

This also fixes #11531 where multiple selinuxfs filesystems could be mounted,
as it returns only the first mountpoint.

The /proc file was changed from /proc/self/mountinfo to /proc/self/mounts for
compatibility with Linux 2.6.25 and older.