(#2157) Ensure externals always overwrite other facts
Before this commit, external facts were not being loaded when a
specific fact was asked for. This meant that asking for all facts
would return different results than asking for an individual fact.

Now, all requests for facts load all exeternal facts, and external
facts have a set weight that is very high.