(#2157) Allow weight to be set during fact addition
Previously the Facter.add() method took parameters that would be used
to call attribute setters on the created resolution. This worked
for the value of the resolution, but not for the weight. The weight
was set after the resolution had been added to the fact, but the
weight ordered facts (for deciding which resolution takes precedence)
was calculated at the time the resolution was added. This meant that
a weight set via an option to the add() method was ignored.

This changes the sort to occur at the time the fact value is evaluated
so that the weight set as an option via add() is taken into account.