(#2157) Prevent external facts from attempting to parse directories
Prior to this commit, external facts would attempt to execute a
file if it was executable, which mean it would attempt to parse
sub directories in the external fact directory it was loading from.
Now it checks to make sure that what it is trying to parse is
executable and is a file, which fixes this issue.