(#2157) Rename collection loader methods
Without this patch the methods used to load facts do not match up with
the instance variables bound to the loaders themselves.

This patch simply renames the methods to match the internal and external
loader instance variable names.  This fixes the problem by making it
explicit which loader is being used when the methods are called on the
collection instance.

Paired-with: Hailee Kenney <hailee@puppetlabs.com>