(#2157) Make JsonParser always available but disabled without json gem
Without this patch the tests for the JsonParser class and the class
itself are not defined if the JSON gem is not available.  This is a
problem because the end user will consider this behavior a "silent
failure" if they place *.json files in the data directory and Facter
silently skips over them because the JSON library is not loadable.

This patch fixes the problem by always defining the JsonParser class.
The class itself will issue a warning when it tries to return results
for a JSON data file and the JSON library is not available.  This patch
also marks the json parser tests as pending if the json library is
unloadable.

This patch also fixes this error with command line option parsing:

    $ facter --json
    invalid option: --json

With this patch we're getting:

    $ facter --json
    You do not have JSON support in your version of Ruby. JSON output disabled

Paired-with: Hailee Kenney <hailee@puppetlabs.com>