(#2157) Make parser matcher work with Enumerable
Without this patch the parser matcher for filename extensions was locked
into the specific class of Array and String.  The implementation is
generally applicable to Enumerable and descendants of String though.

This patch simply changes the conditional checks to use the Ruby idiom
of comparing an object instances to parent classes using a case
statement.  This is equivalent to using kind_of? or is_a? which are
themselves synonyms.

diff --git a/lib/facter/util/parser.rb b/lib/facter/util/parser.rb
index 653f297..05dec2f 100644
--- a/lib/facter/util/parser.rb
+++ b/lib/facter/util/parser.rb
@@ -10,10 +10,11 @@ module Facter::Util::Parser
   # For support mutliple extensions you can pass an array of extensions as
   # +ext+.
   def self.extension_matches?(filename, ext)
-    if ext.class == String then
-      extension = ext.downcase
-    elsif ext.class == Array then
-      extension = ext.collect {|x| x.downcase }
+    extension = case ext
+    when String
+      ext.downcase
+    when Enumerable
+      ext.collect {|x| x.downcase }
     end
     [extension].flatten.to_a.include?(file_extension(filename).downcase)
   end