(#2157) Change parser specs to not touch the filesystem
Without this patch all of the parser specification examples would create
files in the filesystem and read them back out or execute them.

This is a problem because the tests are much slower to run, much harder
to maintain, and there's crap left all over the place on my system.

This patch fixes the problem by changing the parser implementations to
load content on demand.  The spec tests themselves have been changed to
stub out the File.read method when appropriate.  We've also changed the
specifications to stub out Facter::Util::Resolution.exec

Paired-with: Hailee Kenney <hailee@puppetlabs.com>