(#2157) Fixup yaml typo in json spec test
The test for the external fact loading of json data was incorrectly
using a yaml file extension for the filename.  This was causing failures
because the YAML parser was being used instead of the JSON parser.

This patch uses the proper json file extension.

Paired-with: Hailee Kenney <hailee@puppetlabs.com>