(#2157) Fix spec failures on Ruby 1.8.5
The external facts pull request contained a call to String#lines which
is not a valid method in Ruby 1.8.5.  Without this patch the specs fail
correctly.

This patch fixes the problem by changing String#lines to
String#each_line.  The tests pass on CentOS 5 with Ruby 1.8.5-25el5.

This patch also fixes a missing method error in the directory loader
where we were calling String#start_with? which does not exist in Ruby
1.8.5.  This patch replaces the method call with a regular expression.

Finally, the exception handler in the directory loader has been
tightened up to only catch Errno::ENOENT exceptions.

Paired-with: Hailee Kenney <hailee@puppetlabs.com>