(#2157) Fixup remove executable facts on Windows
Without this patch applied the ScriptParser class is still being used to
try and resolve invalid facts on windows.  This causes the following
spec failure:

    2) Facter::Util::DirectoryLoader when loading facts from disk should ignore files with an extension of 'bak'
       Failure/Error: subject.load(collection)
       NoMethodError:
         undefined method `each_line' for nil:NilClass
       # ./spec/../lib/facter/util/parser.rb:118:in `results'
       # ./spec/../lib/facter/util/directory_loader.rb:61:in `load'
       # ./spec/../lib/facter/util/directory_loader.rb:55:in `each'
       # ./spec/../lib/facter/util/directory_loader.rb:55:in `load'
       # ./spec/unit/util/directory_loader_spec.rb:60

This failure is a result of `File.executable?` returning true on Windows when
the file itself is not actually executable, because it is not a PATHEXT file.

This patch also updates the parser specification examples to ensure the
expected results are returned on Windows.  In particular, script parsers should
be disabled on the windows platform.

Paired-with: Hailee Kenney <hailee@puppetlabs.com>