(#2157) Fix Mocha::StubbingError in loader_spec
Without this patch applied the spec tests for the loader_spec.rb file
are setting a stub on a nil object.  This is a problem because Mocha
0.11.3 will raise an exception catching this common error.

This exception is not raised in Mocha 0.10.4.  Mocha 0.11.3 has added a
feature to raise stubbing errors when trying to stub methods on nil
objects.

Paired-with: Hailee Kenney <hailee@puppetlabs.com>