Delete the "ARP" facts, which are pathologically bad.
The "ARP" facts are pretty awesome, really, except for a tiny handful of
little things that could do with improvement.

For example, they are not really about ARP, they are about detecting if the
current node is an EC2 node or not.  So, the name is kind of wrong.

They depend on the fact that the EC2 happened to communicate with a particular
EC2 metadata service first, before doing anything else, because they just
return the first entry in the ARP table attached to an interface.

They are also not used anywhere in our core - even the other facts that used
to use them stopped, or duplicated their functionality.

All in all, they can go: they have so much wrong with them that saving them is
kind of crazy.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>