Delete Linux-only `get_arp_value` helper.
This was only ever implemented for Linux, and only to support the ARP facts
that were all about detecting if we are EC2 or not.

It also kind of sucks, because by "get ARP value" it means "get the first ARP
entry tied to an interface, because picking some random value based on the
oldest machine you communicated with that hasn't fallen out of the cache yet."

If someone actually comes up with a use for this data (eg: some random ARP
value based on previous network behaviour of other tools we run) then I
suppose they can reintroduce it - but, honestly, I can't imagine why.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>