Avoid shell when reading the ARP table on Linux.
We can avoid the overhead of shelling out to read the ARP table on Linux,
since `/proc/net/arp` contains the exact set of information we want.

This is about an order of magnitude faster (0.15 rather than 1.5 ms), which
isn't all that much speed difference, but this is out of the 1.13 seconds a
complete run takes on my system - so ten hits like this is one percent of the
overall runtime.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>