Avoid external command for reading filesystem type list.
This eliminates the use of cat in favour of plain and ordinary read; despite
the brief RedHat reintroduction of the kernel bug when using select on files
in /proc, this isn't worth working around in some-but-not-all places.

It also eliminates most of the lines while still having the same effect,
making the code much cleaner.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>