(#15687) Selinux: Test for policyvers before reading it
Previously facter would read /#{selinux_mount_point}/policyvers without first
verifying it existed, which would spew stderr to the console if it did not
exist. This commit makes the default value for the fact "unknown" and only uses
a different value if policyvers exists. This also includes an updated test
which fails using the previous fact definition.