Add 'operatingsystemmajorrelease' fact on RedHat/Debian
Addition of new fact based on operatingsystemrelease fact.
Whilst this information is available as the lsb majorversion
the dependencies on LSB are high (pulls in graphics and printing)
for minimal installs.

Deliberately constrained to known useful OS releases - ie it would
make no sense to return the 1st part of a Ubuntu release