Add more packaging to rake tasks
This commit adds rpm and debian packaging
for building facter packages. Much of the
packaging functionality is from work done by
Matthaus Litteken for Hiera packaging. rpm
building uses the rpmbuild command, and
pdebuild and a COW setup with pbuilder
are required for building debs. debian building
also requires various build artifacts, which
are added in a separate commit.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>