Add debian build artifacts to facter project
This commit takes the debian build artifacts
from facter-debian-packaging and places them
into facter proper. This is a substantial change
that allows us to stop maintaining the second
repository altogether. The debian changelog is
also moved to an erb file for automation, but
this can still be updated manually if desired.
gbp.conf is not carried over because debs are
build with pdebuild instead of git-build-package.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>