(#15514) Add compatibility with change to operatingsystem fact
Since the operatingsystem fact can now return several different
version of Solaris, anything that relies on operatingsystem returning
Solaris must now depend on osfamily returning Solaris.

Most of these cases have been taken care of, so far this is the
only one I have found that was missed, but it's possible there
are a few more.

<hailee@puppetlabs.com>