Add build-requires of ruby-rdoc for manpage generation
Without a BuildRequires of ruby-rdoc, if we remove the --no-rdoc flag from
install.rb, the manpages will still not be generated, which will cause the
build to fail, because the %files section lists manpages. Adding a
BuildRequires of ruby-rdoc and removing the --no-rdoc flag from install.rb will
allow manpage generation for packaging.