(#15852) Allow RC in Puppet, Facter version strings
Previously putting rc in the Facter version string would fail, mainly because
of a test that verified that the version only contained 3 dot separated
integers. This commit updates the Facter version to correctly use 2.0.0-rc4. It
also removes the test that verified the Facter version was a 'valid' 3 dotted
integer version.