Fix #7559 - Add a flag file to check for presence of EC2
Add a check to Facter to determine if a host is an EC2 or VPC
instance. The current checks for MAC addresses and ARP entries is
not sufficient because Amazon VPC instances do not have known MAC
addresses.

The check assumes a provisioning process or a tool like Cloud
Provisioner populates the file on the host. The check looks for
a file at /etc/facter_ec2 on Unix-like systems and·in the COMMON_APPDATA
directory on Windows systems.