(#15464) Make Facter.version settable via Facter.version=
Without this patch applied the version string of Facter is dynamically
set from the `git describe` tag in the Rakefile using monkey patching.

This is problematic because this causes a constant to be re-defined
which issues a warning in Ruby.

This patch fixes the problem by implementing the Facter.version= module
method.  The Rakefile will now set the Facter version by using `require
'puppet/version'` and calling `Facter.version = %x{git describe ...}`

Using a module method also has the benefit of making it easier to
intercept and mock Facter version specific behavior.

From this point forward, the method Facter.version should but used to
obtain the version string instead of the constant Facter::FACTERVERSION.