Fixup redhat spec file for f17
This commit fixes up the redhat facter.spec file,
adding in some missing fedora 17 changes that were
probably lost in a merge-up. The os-dependent
definition of facter_libdir is replaced, and the
mandir file-listing is added. The last changelog
entry is removed, which is out of chronological
order (which prevents rpm building). It also removes
an errant conflict marker.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>