(#16397) Update selinux_mount_point for 1.8.5 compatibility
The lines method does not exist for the String class in ruby 1.8.5, which
causes the selinux_mount_point function to print Could not retrieve selinux:
undefined method lines' for #String:0x2abc1ffefe88to STDERR. This commit
monkey patches facter with `lines` so that it can safely use the `lines`
method on 1.8.5. It monkey_patches both String and IO.

It also includes spec tests for the String and IO classes to verify they
respond to lines in the expected manner.