(#5205) Update facter manpage
This commit doesn't handle the request in the ticket of dynamically generated
manpage, but it does update the man page to have updated flags and help,
correct license, copyright holder and copyright year. It also fills in the name
of the program for the manpage so that the header doesn't look quite so
mysterious.