Update facter debian packaging for 1.9 support
Because facter works on ruby 1.8 and 1.9, putting it in the vendordir instead
of the vendorlibdir or rubylibdir is preferred so that either ruby can be used
with facter. This updates the dependencies to bring in ruby instead of a
specific ruby version and also replaces ruby1.8 calls in rules with ruby calls.