(maint) Update debian/rules to use new --ruby flag
This updates the install.rb call in debian/rules to use the new --ruby flag and
sets the ruby binary location to /usr/bin/ruby, as is the standard on
debian/ubuntu installations. This is preferred to using env ruby as is in the
facter binary because that depends on the users' path, and might not bring in
the expected ruby.