Remove CHANGELOG from Facter
This commit removes the CHANGELOG file in the top level
of facter, which only contains commit hashes/ messages
from the previous releases, and is of debatable utility.
This information can be gleaned easily from a git log,
and is not of much use without git. Removing it removes
yet another manual process from the releasing of facter.

It also removes references of the CHANGELOG from the packaging.