(#16527) Fix processorcount fact on Solaris
Without this patch applied the processorcount fact on Solaris releases
without the kstat command produces the following error message:

    Could not retrieve processorcount: private method `scan' called for nil:NilClass

This is a problem because there is no clear path forward for the end
user.

This patch fixes the problem by switching the command executed to gather
the processor count from executing `kstat` to executing `psrinfo` for
Solaris kernel releases before 5.8.