(maint) Remove bad test of manufacturer fact on Solaris
Without this patch, the following test is failing when run in a Solaris
10 virtual machine:

  1) Facter::Manufacturer should not set manufacturer or productname if prtdiag output is nil
     Failure/Error: Facter.value(:manufacturer).should be_nil
       expected: nil
            got: "VMware, Inc."
     # ./spec/unit/util/manufacturer_spec.rb:40

This problem is self-evident and is caused by the fact that the test is simply
wrong.  Other implementations of the manufacturer spec may provide a valid
result, which is exactly what is happening in this situation.  In the situation
where `prtdiag` is nil, the manufacturer of the machine appears to be VMware,
which is expected behavior.