(#10278) Add Facter::Util::Resolution.locale_wrapper
Without this patch, Facter does not have a clean way to reset the locale
when executing system commands.  This is a problem because we cannot
depend on the output of commands like `ifconfig` unless we can also
depend on the localization being  used.

This patch addresses the problem by adding the method locale_wrapper
method.  This method sets the LANG, LC_ALL, and LC_CTYPE environment
variables to the value of C before evaluating the given block.