(#10278) Force ifconfig fact to use Facter default locale
Without this patch Facter executes the ifconfig command without
controlling the localization environment.  This is a problem because the
output of ifconfig causes the parser to fail when in the `de_DE.UTF8`
locale.

This patch fixes the problem by switching the ipaddress fact to use the
locale_wrapper method.