Maint: fix mismatched stub
Previously, the infiniband spec test would fail on systems where the
file `/sys/class/net/ib0/address` did not exist, but `/sbin/ip` did, as
it would attempt to execute `/sbin/ip link show ib0`. This change was
introduced in commit 1d82907b.

The real issue was that the code was using `File::exist?`, but the spec
test was stubbing `File.exists?`. And this wasn't noticed in CI, because
jenkins was executing `rake spec`, which doesn't do anything and just
returns 0.

This commit changes the code to use `File.exists?` as that is more
standard usage, and updates the spec test to add expectations on the
correct methods.