(#16511) Do not call arp -an on Solaris nodes
Without this patch applied the EC2 facts cause Facter to execute the
`arp -an` system command on Solaris systems.  This is an invalid command
on Solaris 8, but not 9 or 10.  This is a problem because the following
output is displayed to the user every run:

    arp: -an: unknown host

The other problem is that facts which depend on this behavior working
are themselves broken and do not resolve on Solaris.

This patch _papers over the problem_, but does not fix the core issues
present in the EC2 facts.  The core problem is that way too much EC2
specific work is being done in the main scope, causing this behavior and
code paths to be exercised on systems like Solaris 8 that will never run
in EC2.

The example `arp -a` command example is copied from
http://docstore.mik.ua/orelly/networking_2ndEd/tcp/ch13_04.htm