(#11612) Add support for processorX facts on HP-UX
Without this patch applied, the processorX facts are not available on
HP-UX platforms.  This is a problem because the lack of these facts make
it difficult to write Puppet manifests that work on HP-UX and other
platforms like Linux that support the processorX facts.

These methods involve using machinfo when it is available, using
sched.models and getconf when machinfo is not available.  A suite of
RSpec tests are included as well as fixture files for four selected
examples of machinfo outputs, a sample sched.models file and unistd.h.