Remove hardcoded version from source in facter
This commit removes the hardcoded version in the facter source
code. This allows us to lay down the version file in packaging
while avoiding a commit to source prior to tagging a release.
Once a release passes tests, it should be tagged and shipped,
without needing an extra commit below the tag.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>