(Maint) Properly mock SunOS in test for manufacturer
Previously this test did not properly mock the kernl and hardwareisa
facts, which caused it to actually run the code on solaris. Having the
code actually run on solaris made it return an unexpected value and fail
the test.

This changes the mocking and should cause the test to start passing on
solaris.