(#17493) Safer handling of EC2 'open' calls
Previously, when facter (and therefore) facter was run on an openstack
node, the metadata method did not protect against the 'open' method from
raising exceptions. Since we don't know ahead of time what the name of
the metadata ec2 fact(s) are, we have to add a rescue block outside of
the setcode block. If an exception occurs, we warn, similar to how
facter's built-in resolution mechanism warns about exceptions.

Previously, the ec2_userdata fact performed the 'open' call outside of
the setcode block, and only rescued one type of exception. This commit
moves the 'open' call into the setcode block, which automatically
rescues and warns if needed.

As a result of this change, the ec2_userdata fact is not evaluated until
actually requested, which simplifies the tests for the ec2 metadata
fact(s).

Based on code from Eric Sorenson <eric0@puppetlabs.com>