(#17612) Add Google Compute Engine support to virtual fact
Without this patch the virtual fact does not return a value that
specifically identifies Google Compute Engine.  This is a problem
because configuration decisions may depend on if the resources being
configured exist within GCE or not.

This patch addresses the problem by changing the behavior of the
`virtual` fact to return the value `gce` when running in Google Compute
Engine's cloud.

This patch uses the Google recommended and documented way to detect if
we are running on a Google Compute Engine:
https://developers.google.com/compute/docs/instances#dmi