(maint) Remove hard dependency on rspec from Rakefile
Rspec is currently required to run any rake tasks in Facter, or even to list
the available rake tasks. For users wishing to just build packages, this is a
high bar. This commit moves the rspec and rubygem requires into a begin/rescue
block so that loaderrors can be caught and ignored. It also moves the tasks
that are defined using Rspec::Core::RakeTask into a conditional on whether
Rspec::Core::RakeTask is available to use. This allows the rake tasks to be
listed and run without rspec installed.